requirements.txt
streamlit
python-dotenv
google-generativeai
gTTS
pydub
googletrans==4.0.0-rc1
streamlit-mic-recorder
openai-whisper
torch